move AlignmentInspector to containers
